[Modify] 돈 관련 타입 BigDecimal 로 변경

+++ backend/timebank/src/main/kotlin/kookmin/software/capstone2023/timebank/presentation/api/v1/model/BankAccountCreateResponseData.kt
@@ -1,7 +1,9 @@
 package kookmin.software.capstone2023.timebank.presentation.api.v1.bank.model
 
+import java.math.BigDecimal
+
 data class BankAccountCreateResponseData (
-    val balance: Int,
+    val balance: BigDecimal,
     val accountNumber: String,
     val bankAccountId: Long
 )
